Polish: stuck-measurement rounding parity + procedural icon for launchers
Stuck rounding fix: StuckMeasurement now stores `at` / `start` /
`end` as `f64` instead of rounded `i32`. The renderer's pill text
already does subtract-then-round on `(end - start)`, so it now
matches the live W×H pill exactly. Previously each endpoint was
rounded individually inside `freeze_axis_measurement`, which on
HiDPI could drift the displayed length by a full logical pixel
relative to the live readout (e.g. live 7px → frozen 6px).
Session save/load adapt to floats automatically through
`str::parse` inference.

Desktop-entry icon: the procedural purple/teal app icon is now
rendered at a caller-specified size (the tray still gets 64×64,
the desktop entry gets 256×256) and dropped to
`~/.local/share/icons/hicolor/256x256/apps/vernier.png` on
daemon startup. Both the autostart and application desktop files
now reference the absolute PNG path so launchers (walker, rofi,
GNOME activities) show the same purple icon as the tray instead
of the launcher-theme fallback that looks like generic blue
dots.

+++ crates/pixelsnap-platform/src/types.rs
@@ -245,7 +245,10 @@ pub struct HeldRect {
 
 /// A frozen single-axis measurement. Drawn as a coral line spanning
 /// `start..end` with tick caps on both ends and a pill showing the
-/// pixel length.
+/// pixel length. Positions are kept as `f64` so the renderer can do
+/// the same `subtract-then-round` step the live W×H pill uses; if
+/// the endpoints were rounded individually before subtraction, HiDPI
+/// edge positions could drift the displayed length by 1 px.
 #[derive(Debug, Clone, Copy)]
 pub struct StuckMeasurement {
     /// `Vertical` = a vertical line measuring up-to-down extent.
@@ -253,11 +256,11 @@ pub struct StuckMeasurement {
     pub axis: GuideAxis,
     /// Perpendicular position in logical px (x for Vertical,
     /// y for Horizontal).
-    pub at: i32,
+    pub at: f64,
     /// Start of the measured span in logical px.
-    pub start: i32,
+    pub start: f64,
     /// End of the measured span in logical px.
-    pub end: i32,
+    pub end: f64,
     /// Transient: true when the cursor is over this measurement's
     /// pill — renderer swaps the value text for "×" to signal
     /// "click to remove".
